+++ app/Http/Controllers/User/ArticlesController.php
@@ -9,7 +9,7 @@ use App\Http\Resources\Article\ArticlesResource;
 use App\Http\Resources\User\UserResource;
 use App\Models\Article\Status;
 use App\Models\User\User;
-use App\Service\MarkReactionService;
+use App\Service\FeedbackService;
 use Illuminate\Auth\AuthenticationException;
 use Illuminate\Contracts\Auth\StatefulGuard;
 use Inertia\Inertia;
@@ -20,7 +20,7 @@ class ArticlesController extends Controller
 {
     public function __construct(
         private readonly StatefulGuard $authService,
-        private readonly MarkReactionService $reactionService,
+        private readonly FeedbackService $reactionService,
     ) {
     }
