+++ tests/Unit/Services/FeedbackService/MarkArticleTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\Services\MarkReactionService;
+namespace Tests\Unit\Services\FeedbackService;
 
 use App\Models\Article\Article;
 use App\Models\User\User;
 use App\Repositories\Interfaces\ArticleRepositoryInterface;
 use App\Repositories\Interfaces\CommentRepositoryInterface;
-use App\Service\MarkReactionService;
+use App\Service\FeedbackService;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
@@ -30,7 +30,7 @@ class MarkArticleTest extends TestCase
         $articleRepository = $this->createMock(ArticleRepositoryInterface::class);
         $commentRepository = $this->createMock(CommentRepositoryInterface::class);
 
-        $service = new MarkReactionService($articleRepository, $commentRepository);
+        $service = new FeedbackService($articleRepository, $commentRepository);
 
         $service->markArticle($user, $article);
